moved ccfri home application into its own folder. will break up the form components into seperate sub pages'
