+++ frontend/src/router.js
@@ -21,7 +21,7 @@ import { PATHS } from '@/utils/constants';
 import OrganizationSelection from '@/components/OrganizationSelection';
 
 import CCFRIApplicationForm from '@/components/CCFRIApplicationForm';
-import CcfriEceLandingPage from '@/components/CcfriEceLanding';
+import CcfriEceLandingPage from '@/components/ccfriApplication/group/CcfriEceLanding';
 
 import OrganizationInformation from '@/components/ccofApplication/group/OrganizationInformation';
 import FacilityInformation from '@/components/ccofApplication/group/FacilityInformation';
@@ -165,7 +165,7 @@ const router = new VueRouter({
       }
     },
     {
-      path: '/ccfri-home',
+      path: PATHS.ccfriHome,
       name: 'ccfri-home',
       component: CcfriEceLandingPage,
       meta: {
