+++ frontend/src/utils/constants.js
@@ -35,8 +35,11 @@ export const PATHS = {
   orgInfo: '/ccof-application/group/organization-information',
   facInfo: '/ccof-application/group/facility-information',
   fundAmount: '/ccof-application/group/funding-amount',
-  confirmation: '/ccof-application/group/confirmation'
+  confirmation: '/ccof-application/group/confirmation',
+  ccfriHome: '/ccfriApplication/group/CcfriEceLanding' 
 };
+
+
 
 
 export const MINISTRY_NAME = 'Ministry of Education and Child Care';
